requirements.txt
freqtrade
fastapi
uvicorn
pydantic-settings
sqlalchemy
pandas
psycopg2-binary
alembic
ccxt
joblib
xgboost
lightgbm
torch
scikit-learn